made suggested changes
