+++ feature/update-password/src/commonMain/kotlin/org/mifos/mobile/feature/update/password/UpdatePasswordViewModel.kt
@@ -15,7 +15,12 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import mifos_mobile.feature.update_password.generated.resources.Res
+import mifos_mobile.feature.update_password.generated.resources.confirm_password_error_validation_blank
+import mifos_mobile.feature.update_password.generated.resources.confirm_password_error_validation_minimum_chars
 import mifos_mobile.feature.update_password.generated.resources.could_not_update_password_error
+import mifos_mobile.feature.update_password.generated.resources.error_password_not_match
+import mifos_mobile.feature.update_password.generated.resources.new_password_error_validation_blank
+import mifos_mobile.feature.update_password.generated.resources.new_password_error_validation_minimum_chars
 import org.jetbrains.compose.resources.StringResource
 import org.mifos.mobile.core.common.DataState
 import org.mifos.mobile.core.data.repository.UserAuthRepository
@@ -29,7 +34,7 @@ internal class UpdatePasswordViewModel(
         MutableStateFlow<UpdatePasswordUiState>(UpdatePasswordUiState.Initial)
     val updatePasswordUiState = _updatePasswordUiState.asStateFlow()
 
-    fun updateAccountPassword(newPassword: String, confirmPassword: String) {
+    private fun updateAccountPassword(newPassword: String, confirmPassword: String) {
         viewModelScope.launch {
             when (userAuthRepositoryImp.updateAccountPassword(newPassword, confirmPassword)) {
                 is DataState.Error<*> ->
@@ -44,6 +49,65 @@ internal class UpdatePasswordViewModel(
 //                    clientRepositoryImp.updateAuthenticationToken(newPassword)}
                 }
             }
+        }
+    }
+
+    fun validateAndUpdatePassword(
+        params: PasswordValidationParams,
+    ) {
+        with(params) {
+            val newPasswordErrorContent = getPasswordError(newPassword, PasswordType.NEW)
+            val confirmPasswordErrorContent = getPasswordError(confirmPassword, PasswordType.CONFIRM)
+
+            setNewPasswordErrorContent(newPasswordErrorContent)
+            setConfirmPasswordErrorContent(confirmPasswordErrorContent)
+
+            when {
+                newPasswordErrorContent == null && confirmPasswordErrorContent == null -> {
+                    if (newPassword == confirmPassword) {
+                        updateAccountPassword(newPassword, confirmPassword)
+                    } else {
+                        setPasswordDoesNotMatchOnBothError(Res.string.error_password_not_match)
+                    }
+                }
+
+                newPasswordErrorContent == null && confirmPasswordErrorContent != null -> {
+                    setConfirmPasswordError(true)
+                }
+
+                newPasswordErrorContent != null && confirmPasswordErrorContent == null -> {
+                    setNewPasswordError(true)
+                }
+
+                else -> {
+                    setNewPasswordError(true)
+                    setConfirmPasswordError(true)
+                }
+            }
+        }
+    }
+
+    private enum class PasswordType {
+        NEW,
+        CONFIRM,
+    }
+
+    private fun getPasswordError(
+        password: String,
+        type: PasswordType,
+    ): StringResource? {
+        return when {
+            password.isEmpty() -> when (type) {
+                PasswordType.NEW -> Res.string.new_password_error_validation_blank
+                PasswordType.CONFIRM -> Res.string.confirm_password_error_validation_blank
+            }
+
+            password.length < 6 -> when (type) {
+                PasswordType.NEW -> Res.string.new_password_error_validation_minimum_chars
+                PasswordType.CONFIRM -> Res.string.confirm_password_error_validation_minimum_chars
+            }
+
+            else -> null
         }
     }
 }
